improve error handling

+++ src/datamodel/taskData.js
@@ -14,7 +14,6 @@ export const getUserTasks = async (userId,title,dueDate) => {
       }
       if(!title && dueDate){
         const allTasks = await db.select().from(tasks).where(and(eq(tasks.userId, userId), eq(tasks.dueDate, dueDate)));
-        console.log(allTasks[0].title);
         return allTasks;
       }
       const allTasks = await db.select().from(tasks).where(eq(tasks.userId, userId));
@@ -37,24 +36,24 @@ export const editTask = async (taskId,userId,title,description,dueDate,status) =
   try {
     const dateOnly = new Date(dueDate).toISOString().split('T')[0];
     const currentDate = new Date();
-    await db.update(tasks)
+    const editedTask = await db.update(tasks)
     .set({ userId: userId , title: title, description: description, dueDate: dateOnly,status: status,updatedAt: currentDate })
     .where(and(eq(tasks.id, taskId), eq(tasks.userId, userId)))
     .returning({ updatedId: tasks.id });
-    console.log(newTask);
-    return newTask;
+    return editedTask;
   } catch (error) {
     return Promise.reject(error);
   }
 }
 
-export const removeTask = async (taskId) => {
+export const removeTask = async (taskId,userId) => {
  try {
   const deletedTask = await db.delete(tasks)
-  .where(eq(tasks.id, taskId))
-  .returning({"deletedtask":tasks.id});
+  .where(and(eq(tasks.id, taskId), eq(tasks.userId, userId)))
+  .returning({deletedId: tasks.id});
   return deletedTask;
  } catch (error) {
+  console.log(error);
   return Promise.reject(error);
  }
 }
